Refactor API with code splitting
