+++ apps/webapp/api/v1Api.ts
@@ -1,0 +1,17 @@
+import { Hono } from "@hono/hono";
+import { Env } from "./main.ts";
+
+function cleanUrl(url: string) {
+  return url.replace(/\/(index.html)?\/?$/, "");
+}
+
+const app = new Hono<Env>()
+  .post("/headlines", async (c) => {
+    const { url } = await c.req.json();
+    const dbClient = c.var.db;
+
+    const headlineData = await dbClient.getAllCreatorEdits(cleanUrl(url));
+    return c.json(headlineData);
+  });
+
+export default app;
